updated 1-concurrent_coroutines.py with annotation implementation

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -6,7 +6,7 @@ from typing import List
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: float) -> List[float]:
+async def wait_n(n: int, max_delay: int) -> List[float]:
     """module 2"""
     delays: List[float] = []
     tasks: float = [wait_random(max_delay) for _ in range(n)]
